Change WlanReadWpa return type to void
Arduino core for the ESP32 sets -Werror=all for warning levels of "More" or "All" in the Arduino IDE's File > Preferences > Compiler warnings setting. This upgrades what would have previously been a warning about the lack of a return statement in non-void WlanReadWpa() to a compilation error. Since the call to WlanReadWpa() doesn't expect a return value and there is nothing obvious in the function to return I have changed the return type to void to fix this error/warning.